- fixed comments - added unit test for MTJD - include

+++ src/core/MTJD/manager.h
@@ -51,7 +51,7 @@ namespace Lux
 
 			uint32_t getAffinityMask(uint32_t idx) const;
 
-			JobsTable		m_ready_to_execute[Priority::Count];
+			JobsTable		m_ready_to_execute[(size_t)Priority::Count];
 			JobTransQueue	m_trans_queue;
 			TransTable		m_pending_trans;
 			WorkerTask*		m_worker_tasks;
